% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_adj_matrix.R
\name{cal_adj_matrix}
\alias{cal_adj_matrix}
\title{Title}
\usage{
cal_adj_matrix(sc_data, corr_p_threshold)
}
\arguments{
\item{sc_data}{single-cell RNA-seq expression matrix of related disease. Each row represents a gene and each column represents a sample.}

\item{corr_p_threshold}{The P-value threshold for calculating the adjacency matrix.}
}
\value{
Adjacency matrix for single-cell data
}
\description{
Title
}
\examples{
test=cal_adj_matrix(case1_sc_data_1to100,0.05)
}
