% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_cell_phenotype_labels.R
\name{cal_cell_phenotype_labels}
\alias{cal_cell_phenotype_labels}
\title{Title}
\usage{
cal_cell_phenotype_labels(p.value, p.value_threshold)
}
\arguments{
\item{p.value}{The p-value of the perturbation test}

\item{p.value_threshold}{The threshold for predicting cells with phenotype 1.}
}
\value{
Predicted cell phenotypic labels
}
\description{
Title
}
