% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvalue.R
\name{calculate_pvalue}
\alias{calculate_pvalue}
\title{Title}
\usage{
calculate_pvalue(cells_scores, permuted_cells_scores_matrix, times)
}
\arguments{
\item{cells_scores}{the proximity between cells and the phenotype of interest.}

\item{permuted_cells_scores_matrix}{perturbation results.}

\item{times}{Number of perturbations.}
}
\value{
p value of proximity.
}
\description{
Title
}
