% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ranked.matrix.R
\name{calculate_ranked.matrix}
\alias{calculate_ranked.matrix}
\title{Title}
\usage{
calculate_ranked.matrix(matrix)
}
\arguments{
\item{matrix}{The original expression matrix.}
}
\value{
rank-based matrix.
}
\description{
Title
}
