% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signature.R
\name{calculate_signature}
\alias{calculate_signature}
\title{Title}
\usage{
calculate_signature(ranked.matrix, signature_threshold)
}
\arguments{
\item{ranked.matrix}{rank-based matrix}

\item{signature_threshold}{the size of signature}
}
\value{
gene signature
}
\description{
Title
}
