% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case1_sample_pheno_labels_200to300}
\alias{case1_sample_pheno_labels_200to300}
\title{sample lables}
\format{
asdfasf
}
\usage{
case1_sample_pheno_labels_200to300
}
\description{
sample lables
}
\keyword{datasets}
