% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_to_phenotype.R
\name{cell_to_phenotype}
\alias{cell_to_phenotype}
\title{Title}
\usage{
cell_to_phenotype(
  graph,
  distance_matrix,
  cell_signature,
  phenotye_sample_signature.list
)
}
\arguments{
\item{graph}{PPI network}

\item{distance_matrix}{matrix calculated by igraph}

\item{cell_signature}{cell signature}

\item{phenotye_sample_signature.list}{signatures of Samples belonging to the same phenotype.}
}
\value{
proximity between a cell and the phenotype.
}
\description{
Title
}
