% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_to_sample.R
\name{cell_to_sample}
\alias{cell_to_sample}
\title{Title}
\usage{
cell_to_sample(graph, distance_matrix, cell_signature, sample_signature)
}
\arguments{
\item{graph}{PPI data}

\item{distance_matrix}{matrix calculated by igraph}

\item{cell_signature}{cell signature}

\item{sample_signature}{sample signature}
}
\value{
proximity between a cell and a sample
}
\description{
Title
}
