% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_max_connected_subgraph.R
\name{Extract_max_connected_subgraph}
\alias{Extract_max_connected_subgraph}
\title{Title}
\usage{
Extract_max_connected_subgraph(ppi_data, sc_data, corr_p_threshold)
}
\arguments{
\item{ppi_data}{Network data}

\item{sc_data}{single-cell RNA-seq expression matrix}

\item{corr_p_threshold}{he P-value threshold for calculating the adjacency matrix}
}
\value{
the largest connected subgraph of the input network data
}
\description{
Title
}
