% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotye_sample_signature.R
\name{extract_phenotye_sample_signature}
\alias{extract_phenotye_sample_signature}
\title{Title}
\usage{
extract_phenotye_sample_signature(sample_signature.list, phenotye_labels)
}
\arguments{
\item{sample_signature.list}{all sample signatures}

\item{phenotye_labels}{pheontype labels}
}
\value{
signatures of Samples belonging to the same phenotype
}
\description{
Title
}
