% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_sc.data.R
\name{impute_sc.data}
\alias{impute_sc.data}
\title{Title}
\usage{
impute_sc.data(sc_data)
}
\arguments{
\item{sc_data}{single-cell RNA-seq expression matrix of related disease.}
}
\value{
imputed single-cell matrix
}
\description{
Title
}
