% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACSI.R
\name{PACSI}
\alias{PACSI}
\title{Title}
\usage{
PACSI(
  bulk_data,
  phenotye_labels,
  sc_data,
  ppi_data = NULL,
  sc_VariableFeatures = 2000,
  times = 100,
  ncores = 4,
  spatial = F,
  impute = F,
  bulk_VariableFeatures = F,
  sample_signature_threshold = 150,
  cell_signature_threshold = 150,
  corr_p_threshold = 0.05,
  cell_population_FDR_threshold = 0.05
)
}
\arguments{
\item{bulk_data}{Bulk expression matrix of related disease. Each row represents a gene and each column represents a sample.}

\item{phenotye_labels}{Phenotype annotation of each bulk sample. It should be a binary variable.}

\item{sc_data}{single-cell RNA-seq expression matrix of related disease. Each row represents a gene and each column represents a sample.}

\item{ppi_data}{Network data that fits the input of the igraph package}

\item{sc_VariableFeatures}{The number of highly variable genes that need to be retained during single-cell data preprocessing.}

\item{times}{Number of perturbations.}

\item{ncores}{Number of CPU cores for parallel operations.}

\item{spatial}{If the input of sc_data is a seurat object for the spatial transcriptome, let spatial =T.}

\item{impute}{If you want to fill in missing values in single-cell data, let impute =T.}

\item{bulk_VariableFeatures}{If you need to extract the highly variable genes of the bulk matrix, let bulk_VariableFeatures be the number of features such as bulk_VariableFeatures = 2000.}

\item{sample_signature_threshold}{The size of the sample signature.}

\item{cell_signature_threshold}{The size of the cell signature.}

\item{corr_p_threshold}{The P-value threshold for calculating the adjacency matrix.}

\item{cell_population_FDR_threshold}{The threshold for predicting cells with phenotype 1.}
}
\value{
asdfas
}
\description{
binary group indicator vector,
}
