% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_cell_signature.R
\name{permute_cell_signature}
\alias{permute_cell_signature}
\title{Title}
\usage{
permute_cell_signature(graph, cell_uniprot_signature.list)
}
\arguments{
\item{graph}{PPI data}

\item{cell_uniprot_signature.list}{all cell signatures}
}
\value{
random cell signatures
}
\description{
Title
}
