% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.list_screened.R
\name{signature.list_screened}
\alias{signature.list_screened}
\title{Title}
\usage{
signature.list_screened(uniprot_signature.list, graph)
}
\arguments{
\item{uniprot_signature.list}{uniprot signature}

\item{graph}{network}
}
\value{
IDs that exists on the network.
}
\description{
Title
}
